Moved Gyro methods to Gyro class
